move html routes out of API folder
