+++ server/controllers/api/encounterRoutes.js
