+++ server/controllers/scoreRoutes.js
